+++ Docs/ReleaseNotes.md
@@ -10,6 +10,7 @@ Note that the hardware framework projects have their own release cycle and relea
 - Adding transformation warning for too large arrays and displaying such arrays in the Consumer sample.
 - Adding the ability to configure methods to be inlined without an attribute.
 - Adding option to enable or disable constant substitution.
+- New sample, see `MonteCarloPiEstimator`: A Monte Carlo simulation computing the value of Pi in a highly parallelized manner.
 - New linear-feedback shift register pseudo random number generator: `RandomLfsr` (check out the `MonteCarloPiEstimator` sample on how it can be useful). This is a simpler PRNG than the `RandomMwc64X` implementation we had before solely, and has lower resource usage, but produces lower quality random numbers. Also included the 16b `RandomLfsr16` and a xorshift version, `RandomXorshiftLfsr16`.
 - Implement workaround for an [ILSpy bug](https://github.com/icsharpcode/ILSpy/issues/807).
 - Various smaller bugfixes and improvements.