+++ Docs/ReleaseNotes.md
@@ -8,16 +8,23 @@ Note that the hardware framework projects have their own release cycle and relea
 ## vNext
 
 - Adding transformation warning for too large arrays and displaying such arrays in the Consumer sample.
+- Adding the ability to configure methods to be inlined without an attribute.
+- Adding option to enable or disable constant substitution.
+- New sample, see `MonteCarloPiEstimator`: A Monte Carlo simulation computing the value of Pi in a highly parallelized manner.
+- New linear-feedback shift register pseudo random number generator: `RandomLfsr` (check out the `MonteCarloPiEstimator` sample on how it can be useful). This is a simpler PRNG than the `RandomMwc64X` implementation we had before solely, and has lower resource usage, but produces lower quality random numbers. Also included the 16b `RandomLfsr16` and a xorshift version, `RandomXorshiftLfsr16`.
+- Implement workaround for an [ILSpy bug](https://github.com/icsharpcode/ILSpy/issues/807).
+- Fixing SimdCalculator hardware execution, see the [issue](https://github.com/Lombiq/Hastlayer-SDK/issues/11).
+- Various smaller bugfixes and improvements.
 
 
-## 1.0.10, 08.06.2017
+## 1.0.10, 08.06.2018
 
 - Updating and fixing hardware timing values, making hardware execution more reliable, but in certain cases slightly slower, however also causing lower FPGA resource usage.
 - Fixing that hardware description caching didn't work with certain programs.
 - Improved documentation.
 
 
-## 1.0.9, 18.03.2017
+## 1.0.9, 18.03.2018
 
 - New Loopback sample to test FPGA connectivity and Hastlayer Hardware Framework resource usage.
 - Fixing that binary operator expressions (like  `1 + 2`) calculated with wrongly sized number types and could cause different results on the hardware than in software.
@@ -29,7 +36,7 @@ Note that the hardware framework projects have their own release cycle and relea
 Note that running Hastlayer now requires Visual Studio 2017 or greater (any edition will work).
 
 
-## 1.0.8, 13.01.2017
+## 1.0.8, 13.01.2018
 
 - Adding ability to inline methods to vastly improve performance if the method is small but called a lot of times.
 - Adding support for `ref` and `out` parameters, see the [issue](https://github.com/Lombiq/Hastlayer-SDK/issues/15).